pyproject.toml
[project]
name = "healthyselfjournal"
version = "0.1.0"
description = "Voice journaling CLI with Whisper transcription and Claude follow-ups"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.34.0",
    "httpx>=0.27.2",
    "gjdutils",
    "jinja2>=3.1.2",
    "numpy>=1.26",
    "openai>=1.44.0",
    "python-dotenv>=1.0.1",
    "questionary>=2.0.1",
    "pyyaml>=6.0.1",
    "readchar>=4.0.6",
    "rich>=13.7.0",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "typer>=0.9.0",
]

[project.scripts]
healthyselfjournal = "healthyselfjournal.__main__:app"

[tool.uv]
package = true
dev-dependencies = [
    "pytest>=8.3.2",
]

[tool.uv.sources]
gjdutils = { path = "./gjdutils", editable = true }

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["healthyselfjournal*"]
